+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for SQL queries with comments to `DuckDB` source
 - Added "WITH" to query keywords in `DuckDB` source
 
+### Changed
+- Changed `duckdb` version to `0.5.1`
+
+### Fixed
+- Fixed `test_duckdb_to_sql_server.py` tests - revert to a previous version
+- Removed `test__check_if_schema_exists()` test
+
+
 ## [0.4.9] - 2022-09-27
 ### Added
 - Added new column named `_viadot_downloaded_at_utc` in genesys files with the datetime when it is created.
@@ -20,9 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new flow file `mindful_to_adls` to upload data from Mindful API tp ADLS.
 - Added `recursive` parameter to `AzureDataLakeList` task
 
+
 ## [0.4.8] - 2022-09-06
 ### Added
 - Added `protobuf` library to requirements
+
 
 ## [0.4.7] - 2022-09-06
 ### Added
@@ -83,6 +93,7 @@ DF to string before adding metadata
 - Added `if_empty` parameter in `DuckDBCreateTableFromParquet` task and in `EpicorToDuckDB`, `SAPToDuckDB`,
 `SQLServerToDuckDB` flows to check if output Parquet is empty and handle it properly.
 - Added `check_if_empty_file()` and `handle_if_empty_file()` in `utils.py`
+
 
 ## [0.4.4] - 2022-06-09
 ### Added